Refactors project structure for constants and prompts
Moves constants and prompts to a dedicated utils subdirectory for better organization and clarity.

This change improves code maintainability and reduces the potential for circular dependencies by centralizing configuration elements.
