+++ audify/translate.py
@@ -3,12 +3,12 @@ from typing import Optional
 
 from langchain_ollama import OllamaLLM
 
-from audify.constants import (
+from audify.utils.constants import (
     LANGUAGE_NAMES,
     OLLAMA_API_BASE_URL,
     OLLAMA_DEFAULT_TRANSLATION_MODEL,
 )
-from audify.prompts import TRANSLATE_PROMPT
+from audify.utils.prompts import TRANSLATE_PROMPT
 
 # Configure logging
 logger = logging.getLogger(__name__)
